Overhaul importation into Gitlab
Instead of flattening the Bugzilla XML early (which is troublesome for
duplicate tags), just use the XML tree directly in the importation code.
This makes it much easier to properly handle attachments from all
Bugzilla users, instead of just the reporter.

In this way, all attachments are included, and obsolete attachments are
marked with strikethrough text. Additionally, instead of more costly and
hackish parsing of additional Bugzilla API calls to get each
attachment's metadata and content, just use the attachment metadata and
base64 encoded content that's already included in the bug XML.

The output table has been adjusted and documented accordingly.